Update fabric8 kubernetes-client from 4.6.4 to 5.0.2 (#5265)
* Update fabric8 kubernetes-client from 4.6.4 to 5.0.2

https://github.com/fabric8io/kubernetes-client/blob/master/doc/MIGRATION-v5.md

+++ address-space-controller/src/main/java/io/enmasse/controller/AddressFinalizerController.java
@@ -21,7 +21,7 @@ public class AddressFinalizerController extends AbstractFinalizerController {
 
     public static final String FINALIZER_ADDRESSES = "enmasse.io/addresses";
 
-    private static final Integer BATCH_SIZE = Integer.getInteger("io.enmasse.controller.AddressFinalizerController.batchSize", 100);
+    private static final Long BATCH_SIZE = Long.getLong("io.enmasse.controller.AddressFinalizerController.batchSize", 100L);
 
     private AddressSpaceApi addressSpaceApi;
