+++ k8s-api-testutil/src/main/java/io/enmasse/k8s/api/TestAddressApi.java
@@ -79,7 +79,7 @@ public class TestAddressApi implements AddressApi {
     }
 
     @Override
-    public ContinuationResult<Address> listAddresses(String namespace, Integer limit, ContinuationResult<Address> continueValue, Map<String, String> labels) {
+    public ContinuationResult<Address> listAddresses(String namespace, Long limit, ContinuationResult<Address> continueValue, Map<String, String> labels) {
         return ContinuationResult.from(listAddresses(namespace), null);
     }
